@@ -2,22 +2,24 @@
   "name": "glitch-hello-node",
   "version": "0.1.0",
   "description": "A simple Node app built on fastify, instantly up and running. Built to be remixed on Glitch.com.",
+  "type": "module",
   "main": "server.js",
   "scripts": {
     "start": "node server.js"
   },
   "dependencies": {
-    "fastify": "^4.21.0",
-    "handlebars": "^4.7.8",
     "@fastify/formbody": "^7.4.0",
     "@fastify/static": "^6.10.2",
     "@fastify/view": "^8.0.0",
-    "socket.io": "^4.4.0",
+    "fastify": "^4.21.0",
     "fastify-socket.io": "^4.0.0",
+    "firebase": "^11.8.1",
+    "handlebars": "^4.7.8",
+    "socket.io": "^4.4.0",
     "sqlite3": "^5.1.7"
   },
   "engines": {
-    "node": "14.x"
+    "node": "22.x"
   },
   "repository": {
     "url": "https://glitch.com/edit/#!/glitch-hello-node"
