@@ -1,15 +1,32 @@
-/**
- * This is the main Node.js server script for your project
- * Check out the two endpoints this back-end API provides in fastify.get and fastify.post below
- */
+// Import the functions you need from the SDKs you need
+import { initializeApp } from "firebase/app";
 
-const path = require("path");
+// TODO: Add SDKs for Firebase products that you want to use
+// https://firebase.google.com/docs/web/setup#available-libraries
+
+// Your web app's Firebase configuration
+// For Firebase JS SDK v7.20.0 and later, measurementId is optional
+const firebaseConfig = {
+  apiKey: "AIzaSyD0Kr5GAyn8dryNSYPiYogX9Xwe3BnGf5E",
+  authDomain: "match-made-on-stage.firebaseapp.com",
+  projectId: "match-made-on-stage",
+  storageBucket: "match-made-on-stage.firebasestorage.app",
+  messagingSenderId: "43496683708",
+  appId: "1:43496683708:web:d8c87f15d58b4b578778b9",
+  measurementId: "G-P9BSKCXKMC"
+};
+
+// Initialize Firebase
+const firebaseApp = initializeApp(firebaseConfig);
+
+import path from "path";
+import { fileURLToPath } from 'url';
 
 // init sqlite db
-const fs = require("fs");
+import fs from "fs";
 const dbFile = "./.data/votes.db";
 const exists = fs.existsSync(dbFile);
-const sqlite3 = require("sqlite3").verbose();
+import sqlite3 from "sqlite3";
 const db = new sqlite3.Database(dbFile);
 
 // if ./.data/votes.db does not exist, create it, otherwise print recent records to console
@@ -59,41 +76,50 @@ const updateVotes = (votes) => {
 }
 
 // Require the fastify framework and instantiate it
-const fastify = require("fastify")({
+import { fastify } from "fastify";
+const app = fastify({
   // Set this to true for detailed logging:
-  logger: false,
+  // logger: false,
+  trustProxy: true, // https://fastify.dev/docs/latest/Guides/Serverless/#google-cloud-run
 });
 
-// ADD FAVORITES ARRAY VARIABLE FROM TODO HERE
+import { fastifyStatic } from "@fastify/static";
+import { fastifyFormbody } from "@fastify/formbody";
+import { fastifyView } from "@fastify/view";
+import fastifySocketPkg from "fastify-socket.io";
+import handlebarsPkg from "handlebars";
+
+const __filename = fileURLToPath(import.meta.url); // get the resolved path to the file
+const __dirname = path.dirname(__filename); // get the name of the directory
 
 // Setup our static files
-fastify.register(require("@fastify/static"), {
+app.register(fastifyStatic, {
   root: path.join(__dirname, "public"),
   prefix: "/", // optional: default '/'
 });
 
 // Formbody lets us parse incoming forms
-fastify.register(require("@fastify/formbody"));
+app.register(fastifyFormbody);
 
 // View is a templating manager for fastify
-fastify.register(require("@fastify/view"), {
+app.register(fastifyView, {
   engine: {
-    handlebars: require("handlebars"),
+    handlebars: handlebarsPkg,
   },
 });
 
-fastify.register(require("fastify-socket.io"), {
+app.register(fastifySocketPkg, {
   // put your options here
   cors: {
-    origin: ['https://www.matchmadeonstage.com', 'https://matchmadeonstage.glitch.me']
+    origin: ['https://www.matchmadeonstage.com', 'https://match-made-on-stage-server--match-made-on-stage.us-central1.hosted.app']
   }
 });
 
 // Load and parse SEO data
-const seo = require("./src/seo.json");
-if (seo.url === "glitch-default") {
-  seo.url = `https://${process.env.PROJECT_DOMAIN}.glitch.me`;
-}
+// import seo from './src/seo.json' with { type: 'json' }; 
+// if (seo.url === "glitch-default") {
+//   seo.url = `https://${process.env.PROJECT_DOMAIN}.glitch.me`;
+// }
 
 // votes simply stored as number
 let votes = 0;
@@ -104,20 +130,20 @@ loadVotes(getDateKey()); // load from DB
  *
  * Returns src/pages/index.hbs with data built into it
  */
-fastify.get("/", function (request, reply) {
+app.get("/", function (request, reply) {
   // params is an object we'll pass to our handlebars template
-  let params = { seo: seo };
+  // let params = { seo: seo };
 
   // The Handlebars code will be able to access the parameter values and build them into the page
-  return reply.view("/src/pages/index.hbs", params);
+  return reply.view("/src/pages/index.hbs");
 });
 
-fastify.get("/results", function (request, reply) {
+app.get("/results", function (request, reply) {
   // params is an object we'll pass to our handlebars template
-  let params = { seo: seo };
+  // let params = { seo: seo };
 
   // The Handlebars code will be able to access the parameter values and build them into the page
-  return reply.view("/src/pages/results.hbs", params);
+  return reply.view("/src/pages/results.hbs");
 });
 
 /**
@@ -125,9 +151,9 @@ fastify.get("/results", function (request, reply) {
  *
  * Accepts body data indicating the user choice
  */
-fastify.post("/", function (request, reply) {
+app.post("/", function (request, reply) {
   // Build the params object to pass to the template
-  let params = { seo: seo };
+  // let params = { seo: seo };
 
 //   // If the user submitted a color through the form it'll be passed here in the request body
 //   let color = request.body.color;
@@ -156,15 +182,15 @@ fastify.post("/", function (request, reply) {
 //   }
 
   // The Handlebars template will use the parameter values to update the page with the chosen color
-  return reply.view("/src/pages/index.hbs", params);
+  return reply.view("/src/pages/index.hbs");
 });
 
-fastify.get("/reset", function (request, reply) {
+app.get("/reset", function (request, reply) {
   if (request.query.password === process.env.ADMIN_PASSWORD) {
     votes = 0;
     updateVotes(votes);
     
-    fastify.io.emit('update', 0);
+    app.io.emit('update', 0);
     
     return reply.send({ message: 'Voting successfully reset' });
   }
@@ -172,10 +198,10 @@ fastify.get("/reset", function (request, reply) {
   return reply.send({ message: 'Requires admin password' });
 });
 
-fastify.ready((err) => {
+app.ready((err) => {
   if (err) throw err
 
-  fastify.io.on('connect', (socket) => {
+  app.io.on('connect', (socket) => {
     console.info('Socket connected!', socket.id);
     
     socket.emit('update', votes);
@@ -190,7 +216,7 @@ fastify.ready((err) => {
       if (votes < -10) votes = -10;
       if (votes > 100) votes = 100;
 
-      fastify.io.emit('update', votes);
+      app.io.emit('update', votes);
       
       updateVotes(votes);
     });
@@ -205,9 +231,12 @@ fastify.ready((err) => {
 
 
 // Run the server and report out to the logs
-fastify.listen(
-  { port: process.env.PORT, host: "0.0.0.0" },
-  function (err, address) {
+const IS_GOOGLE_CLOUD_RUN = process.env.K_SERVICE !== undefined;
+const port = IS_GOOGLE_CLOUD_RUN ? 8080 : process.env.PORT || 8080; 
+const host = IS_GOOGLE_CLOUD_RUN ? "0.0.0.0" : undefined;
+app.listen(
+  { port: port, host: host },
+ (err, address) => {
     if (err) {
       console.error(err);
       process.exit(1);
